Add @GroupTitle and @GroupInitialArguments commands
Previously, there was no way to modify the subsection heading of a Group, but
the documentation even stated "Note that it is currently not possible to give
a header to the Group in the manual..." Hence, the addition of a @GroupTitle
command to allow exactly that.

In addition, the functions in a Group often share an initial set of arguments,
and so this adds a @GroupInitialArguments command to allow such arguments to
be defined once for the declarations in a @BeginGroup/@EndGroup block.

Both commands are added to the documentation, with examples in the Grouping
section, and to the general worksheet test.