implement "ignore" argument for @BeginExampleSession and fix two nits
Sometimes (in particular when trying make a file that is both valid MarkDown
and AutoDoc) it is useful to have an extra string that is ignored in an
ExampleSession. This commit provides an argument to @BeginExampleSession that
specifies such a string.

In addition, it adds the missing "info_classes" to the list of types in the
default chapter data, and it recognizes that often the default value for the
return_value of a man item is [ ] rather than false.