Allow following ReadPackage() calls to scan for docs
In some cases it may make the most sense to scan GAP code for AutoDoc
documentation commands in the same order that GAP itself reads the
code, to achieve a more natural order to the documentation. This commit achieves
roughly that when the autodoc.follow_package option is set; the files referenced
by ReadPackage() calls are read in order after the current file but before any
other files that were waiting. (GAP would of course *interrupt* the reading of
the current file when a ReadPackage() call is reached and resume after, but
this commit does not add the complexity of a stack of streams in progress
to the AutoDoc parser, which would be necessary to emulate
that exact behavior.)

In addition, this commit makes the list of suffixes scanned for by
autodoc.scan_dirs into a configurable option, namely autodoc.scan_suffixes. This
enables the settings scan_dirs := ["."], scan_suffixes := ["g"], and
follow_package := true, which should limit AutoDoc to scanning exactly the files
that the package will load. So the self-test (tst/dogfood.tst) is changed to
use this list of options by way of testing the follow_package feature.